Dynamically assign the active nav item.
This change makes the active nav class be assigned to the currently viewed page instead of just be assigned to the first nav item.

It's a little ugly and I'm new to editing Hexo so feel free to implement this better.

If your intention was to have the first nav item simply be featured, then you can disregard this proposal. (but I would recommend you change the class name from "active" to "featured")

Great theme!